Develop

+++ App/Controllers/ModulePhoneBookController.php
@@ -123,6 +123,7 @@ class ModulePhoneBookController extends BaseController
         $parameters['columns'] = [
             'call_id',
             'number' => 'number_rep',
+            'created' => 'created',
             'DT_RowId' => 'id',
         ];
         $parameters['order'] = ['call_id desc'];
@@ -156,20 +157,18 @@ class ModulePhoneBookController extends BaseController
 
         $dataId = $this->request->getPost('id', ['string', 'trim']);
         $callId = $this->request->getPost('call_id', ['string', 'trim']);
-        $number = $this->request->getPost('number', ['alnum']);
-        $numberRep = $this->request->getPost('number_rep', ['string', 'trim'], $number);
+        $numberRep = $this->request->getPost('number_rep', ['string', 'trim']);
+        $number = PhoneBook::cleanPhoneNumber($numberRep, TRUE);
 
         if (empty($callId) || empty($number)) {
             return;
         }
 
         // If we are unable to change the primary field, delete the old record and recreate it
-        $oldId = null;
         $record = null;
         if (stripos($dataId, 'new') === false) {
             $record = PhoneBook::findFirstById($dataId);
-            if ($record->number !== $number) {
-                $oldId = $record->id;
+            if ($record !== null && $record->number !== $number) {
                 $record->delete();
                 $record = null;
             }
@@ -179,39 +178,17 @@ class ModulePhoneBookController extends BaseController
             $record = new PhoneBook();
         }
 
-        foreach ($record as $key => $value) {
-            switch ($key) {
-                case 'id':
-                    break;
-                case 'number':
-                    $record->number = $number;
-                    break;
-                case 'number_rep':
-                    $record->number_rep = $numberRep;
-                    break;
-                case 'call_id':
-                    $record->call_id = $callId;
-                    break;
-                case 'search_index':
-                    // Collect data for the search index
-                    $username = mb_strtolower($callId);
-                    // Combine all fields into a single string
-                    $record->search_index = $username . $number . $numberRep;
-                    break;
-                default:
-                    break;
-            }
-        }
+        $record->setPhonebookRecord($callId, $numberRep);
 
         if ($record->save() === false) {
             $errors = $record->getMessages();
             $this->flash->error(implode('<br>', $errors));
             $this->view->success = false;
-
+            $this->response->setStatusCode(500);
             return;
         }
 
-        $this->view->data = ['oldId' => $oldId, 'newId' => $record->id];
+        $this->view->data = ['oldId' => $dataId, 'newId' => $record->id];
         $this->view->success = true;
     }
 
@@ -236,22 +213,24 @@ class ModulePhoneBookController extends BaseController
      */
     public function deleteAllRecordsAction(): void
     {
-        $records = PhoneBook::find();
-        foreach ($records as $record) {
-            if (!$record->delete()) {
-                $this->flash->error(implode('<br>', $record->getMessages()));
-                $this->view->result = false;
-                return;
-            }
+        $phoneBook = new PhoneBook();
+        $connection = $phoneBook->getWriteConnection();
+        $tableName = $phoneBook->getSource();
+
+        try {
+            $connection->execute("DELETE FROM {$tableName}");
+            $this->view->result = true;
+            $this->view->reload = 'module-phone-book/module-phone-book/index';
+        } catch (\Throwable $e) {
+            $this->flash->error($e->getMessage());
+            $this->view->result = false;
         }
-        $this->view->result = true;
-        $this->view->reload = 'module-phone-book/module-phone-book/index';
     }
 
     /**
-     * Toggle input mask feature.
+     * Save settings
      */
-    public function toggleDisableInputMaskAction(): void
+    public function saveSettingsAction(): void
     {
         if (!$this->request->isPost()) {
             return;
@@ -262,10 +241,19 @@ class ModulePhoneBookController extends BaseController
             $settings = new Settings();
         }
 
-        $settings->disableInputMask = $this->request->getPost('disableInputMask') === 'true' ? '1' : '0';
+        if ($this->request->hasPost('disableInputMask')) {
+            $settings->disableInputMask = $this->request->getPost('disableInputMask') === 'true' ? '1' : '0';
+        }
+
+        if ($this->request->hasPost('phoneBookApiUrl')) {
+            $settings->phoneBookApiUrl = empty($this->request->getPost('phoneBookApiUrl')) ? NULL : $this->request->getPost('phoneBookApiUrl', 'trim');
+            $settings->phoneBookLifeTime = empty($this->request->getPost('phoneBookLifeTime')) ? 0 : $this->request->getPost('phoneBookLifeTime', 'int!');
+        }
+
         if (!$settings->save()) {
             $this->flash->error(implode('<br>', $settings->getMessages()));
             $this->view->success = false;
+            $this->response->setStatusCode(500);
             return;
         }
         $this->view->success = true;
